feat(slots) abstract the slots operations
